encryption: move AWS KMS provider into cloud/aws
add a new create encryption/export
