+++ Makefile
@@ -97,6 +97,11 @@ else
 # Caller is responsible for setting up test engine features
 endif
 
+ifneq ($(NO_CLOUD),1)
+ENABLE_FEATURES += cloud-aws
+ENABLE_FEATURES += cloud-gcp
+endif
+
 PROJECT_DIR:=$(shell dirname $(realpath $(lastword $(MAKEFILE_LIST))))
 
 BIN_PATH = $(CURDIR)/bin
