+++ components/cloud/aws/Cargo.toml
@@ -6,10 +6,19 @@ publish = false
 
 [features]
 default = ["protobuf-codec"]
-protobuf-codec = ["kvproto/protobuf-codec"]
-prost-codec = ["kvproto/prost-codec"]
+protobuf-codec = [
+    "kvproto/protobuf-codec",
+    "encryption/protobuf-codec",
+]
+prost-codec = [
+    "kvproto/prost-codec",
+    "encryption/prost-codec",
+]
 
 [dependencies]
+bytes = "0.5"
+encryption = { path = "../../encryption", default-features = false }
+futures = "0.3"
 futures-util = { version = "0.3", default-features = false, features = ["io"] }
 # This is only a dependency to vendor openssl for rusoto. It's not clear exactly
 # how openssl is built for tikv, but it seems to be controlled by grpcio. This
@@ -21,12 +30,14 @@ hyper-tls = "0.4.1"
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
 rusoto_core = "0.45.0"
 rusoto_credential = "0.45.0"
+rusoto_kms = { version = "0.45.0", features = ["serialize_structs"] }
 rusoto_s3 = "0.45.0"
 rusoto_util = { path = "../../rusoto_util" }
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 # better to not use slog-global, but pass in the logger
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
 external_storage = { path = "../../external_storage" }
+tikv_util = { path = "../../tikv_util", default-features = false }
 tokio = { version = "0.2.13", features = ["time"] }
 url = "2.0"
