+++ components/cloud/aws/src/lib.rs
@@ -3,3 +3,6 @@ extern crate slog_global;
 
 mod s3;
 pub use s3::{Config, S3Storage};
+
+mod kms;
+pub use kms::AwsKms;
