+++ components/encryption/export/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "encryption_export"
+version = "0.0.1"
+edition = "2018"
+publish = false
+
+[features]
+default = ["protobuf-codec", "cloud-aws", "cloud-gcp"]
+protobuf-codec = [
+  "kvproto/protobuf-codec",
+  "encryption/protobuf-codec",
+]
+prost-codec = [
+  "kvproto/prost-codec",
+  "encryption/prost-codec",
+]
+cloud-aws = ["aws"]
+cloud-gcp = ["gcp"]
+
+[dependencies]
+aws = { path = "../../cloud/aws", optional = true }
+gcp = { path = "../../cloud/gcp", optional = true }
+encryption = { path = "../", default-features = false }
+kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
+openssl = "0.10"
+protobuf = "2.8"
+slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
+# better to not use slog-global, but pass in the logger
+slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
+tikv_util = { path = "../../tikv_util", default-features = false }
+
+[dev-dependencies]
+rust-ini = "0.14.0"
+structopt = "0.3"
