+++ components/encryption/src/lib.rs
@@ -25,9 +25,13 @@ mod metrics;
 pub use self::config::*;
 pub use self::crypter::{
     encryption_method_from_db_encryption_method, verify_encryption_config, AesGcmCrypter, Iv,
+    PlainKey,
 };
 pub use self::encrypted_file::EncryptedFile;
 pub use self::errors::{Error, Result};
 pub use self::io::{create_aes_ctr_crypter, DecrypterReader, EncrypterReader, EncrypterWriter};
 pub use self::manager::{DataKeyManager, DataKeyManagerArgs};
-pub use self::master_key::{create_backend, AwsKms, Backend, FileBackend, KmsBackend};
+pub use self::master_key::{
+    Backend, DataKeyPair, EncryptedKey, FileBackend, KeyId, KmsBackend, KmsProvider,
+    PlaintextBackend,
+};
