+++ components/external_storage/export/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2018"
 publish = false
 
 [features]
-default = ["protobuf-codec", "gcp", "aws"]
+default = ["protobuf-codec", "cloud-gcp", "cloud-aws"]
 protobuf-codec = [
   "kvproto/protobuf-codec",
   "external_storage/protobuf-codec",
@@ -18,6 +18,8 @@ prost-codec = [
   "gcp/prost-codec",
   "aws/prost-codec",
 ]
+cloud-aws = ["aws"]
+cloud-gcp = ["gcp"]
 
 [dependencies]
 libloading = { version = "0.5.2", optional = true }
