+++ components/test_util/Cargo.toml
@@ -5,24 +5,26 @@ edition = "2018"
 publish = false
 
 [features]
-default = ["protobuf-codec"]
+default = ["protobuf-codec", "cloud-aws", "cloud-gcp"]
 protobuf-codec = [
-  "encryption/protobuf-codec",
+  "encryption_export/protobuf-codec",
   "grpcio/protobuf-codec",
   "kvproto/protobuf-codec",
   "security/protobuf-codec",
   "tikv_util/protobuf-codec",
 ]
 prost-codec = [
-  "encryption/prost-codec",
+  "encryption_export/prost-codec",
   "grpcio/prost-codec",
   "kvproto/prost-codec",
   "security/prost-codec",
   "tikv_util/prost-codec",
 ]
+cloud-aws = [ "encryption_export/cloud-aws" ]
+cloud-gcp = [ "encryption_export/cloud-gcp" ]
 
 [dependencies]
-encryption = { path = "../encryption", default-features = false }
+encryption_export = { path = "../encryption/export", default-features = false }
 fail = "0.4"
 grpcio = { version = "0.7", default-features = false, features = ["openssl-vendored"] }
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
