+++ src/lib.rs
@@ -52,8 +52,6 @@ extern crate failure;
 #[cfg(test)]
 extern crate test;
 
-extern crate encryption;
-
 pub mod config;
 pub mod coprocessor;
 pub mod import;
